+++ lib/crs_parser/test/test__crs_parser.ml
@@ -25,7 +25,7 @@ module Getters = struct
   type t =
     { path : Vcs.Path_in_repo.t
     ; content : string
-    ; kind : Cr_comment.Kind.t
+    ; status : Cr_comment.Status.t
     ; qualifier : Cr_comment.Qualifier.t option
     ; priority : Cr_comment.Priority.t
     }
@@ -34,7 +34,7 @@ module Getters = struct
   let of_cr cr =
     { path = Cr_comment.path cr
     ; content = Cr_comment.content cr
-    ; kind = Cr_comment.kind cr
+    ; status = Cr_comment.status cr
     ; qualifier =
         (match Cr_comment.header cr with
          | Error _ -> None
@@ -94,7 +94,7 @@ let%expect_test "invalid syntax CR" =
      (getters (
        (path    my_file.ml)
        (content CR)
-       (kind    CR)
+       (status  CR)
        (qualifier ())
        (priority Now))))
     |}];
@@ -119,7 +119,7 @@ let%expect_test "zero spaces CR" =
          (stop  my_file.ml:2:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:2)
@@ -141,7 +141,7 @@ let%expect_test "zero spaces CR" =
      (getters (
        (path    my_file.ml)
        (content "CR user: Hey.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -160,7 +160,7 @@ let%expect_test "zero spaces CR" =
          (stop  my_file.ml:2:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:1)
@@ -182,7 +182,7 @@ let%expect_test "zero spaces CR" =
      (getters (
        (path    my_file.ml)
        (content "CR user: Hey.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -207,7 +207,7 @@ let%expect_test "multiple spaces CR" =
          (stop  my_file.ml:2:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:4)
@@ -229,7 +229,7 @@ let%expect_test "multiple spaces CR" =
      (getters (
        (path    my_file.ml)
        (content "CR user: Blah.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -252,7 +252,7 @@ let%expect_test "empty CR" =
          (stop  my_file.ml:2:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -274,7 +274,7 @@ let%expect_test "empty CR" =
      (getters (
        (path    my_file.ml)
        (content "CR user:")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -308,7 +308,7 @@ let () = ()
          (stop  my_file.ml:1:29)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -330,7 +330,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR-soon user: Some text")
-       (kind    CR)
+       (status  CR)
        (qualifier (Soon))
        (priority Soon))))
     ========================
@@ -342,7 +342,7 @@ let () = ()
          (stop  my_file.ml:3:32)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:3:3)
@@ -364,7 +364,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR-someday user: Some text")
-       (kind    CR)
+       (status  CR)
        (qualifier (Someday))
        (priority Someday))))
     ========================
@@ -376,7 +376,7 @@ let () = ()
          (stop  my_file.ml:5:40)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:5:3)
@@ -402,7 +402,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR-soon user1 for user2: Some text")
-       (kind    CR)
+       (status  CR)
        (qualifier (Soon))
        (priority Soon))))
     ========================
@@ -414,7 +414,7 @@ let () = ()
          (stop  my_file.ml:7:43)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:7:3)
@@ -440,7 +440,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR-someday user1 for user2: Some text")
-       (kind    CR)
+       (status  CR)
        (qualifier (Someday))
        (priority Someday))))
     ========================
@@ -452,7 +452,7 @@ let () = ()
          (stop  my_file.ml:9:30)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:9:3)
@@ -474,7 +474,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "XCR-soon user: Some text")
-       (kind    XCR)
+       (status  XCR)
        (qualifier (Soon))
        (priority Now))))
     ========================
@@ -486,7 +486,7 @@ let () = ()
          (stop  my_file.ml:11:33)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:11:3)
@@ -508,7 +508,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "XCR-someday user: Some text")
-       (kind    XCR)
+       (status  XCR)
        (qualifier (Someday))
        (priority Now))))
     |}];
@@ -552,7 +552,7 @@ let () = ()
        (path my_file.ml)
        (content
         "CR-2026-01-31 user: This CR has a due date, by not correctly specified.")
-       (kind CR)
+       (status CR)
        (qualifier ())
        (priority Now))))
     ========================
@@ -570,7 +570,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR-20260131 user: This is not it either..")
-       (kind    CR)
+       (status  CR)
        (qualifier ())
        (priority Now))))
     ========================
@@ -582,7 +582,7 @@ let () = ()
          (stop  my_file.ml:7:75)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:7:3)
@@ -606,7 +606,7 @@ let () = ()
        (path my_file.ml)
        (content
         "CR-202601 user: You would presumably only include the year and month.")
-       (kind CR)
+       (status CR)
        (qualifier (Someday))
        (priority Someday))))
     |}];
@@ -631,7 +631,7 @@ let () = ()
          (stop  my_file.ml:1:25)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -653,7 +653,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR user: A first CR")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -665,7 +665,7 @@ let () = ()
          (stop  my_file.ml:1:73)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:41)
@@ -687,7 +687,7 @@ let () = ()
      (getters (
        (path    my_file.ml)
        (content "CR user: Followed by another.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -745,7 +745,7 @@ end
        (whole_loc _)
        (header (
          Ok (
-           (kind      CR)
+           (status    CR)
            (qualifier None)
            (reporter  user1)
            (recipient ()))))
@@ -1034,7 +1034,7 @@ span multiple lines too.
          (stop  my_file.ml:1:33)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -1056,7 +1056,7 @@ span multiple lines too.
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1069,7 +1069,7 @@ span multiple lines too.
          (stop  my_file.ml:6:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:3)
@@ -1091,7 +1091,7 @@ span multiple lines too.
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\nspan multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1113,7 +1113,7 @@ span multiple lines too.
          (stop  my_file.ml:1:42)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -1135,7 +1135,7 @@ span multiple lines too.
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a single line comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1148,7 +1148,7 @@ span multiple lines too.
          (stop  my_file.ml:5:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:3)
@@ -1172,7 +1172,7 @@ span multiple lines too.
        (path my_file.ml)
        (content
         "XCR user: This syntax can be used to write\n// comments that span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1210,7 +1210,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:1:29)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:2)
@@ -1232,7 +1232,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1245,7 +1245,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:4:26)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:2)
@@ -1267,7 +1267,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\n# span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1279,7 +1279,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:6:67)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:6:13)
@@ -1301,7 +1301,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may be left next to a non-empty line.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1316,7 +1316,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:12:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:8:4)
@@ -1340,7 +1340,7 @@ Hello text # $CR user: Comment may be left next to a non-empty line.
        (path my_file.ml)
        (content
         "CR user: Let's cover the case\n  #\n  # Where there are empty lines that are part of the comment.\n  # This happens when the CR has multiple paragraphs.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1373,7 +1373,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:1:30)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -1395,7 +1395,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1408,7 +1408,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:4:27)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:3)
@@ -1430,7 +1430,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\n## span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1442,7 +1442,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:6:68)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:6:14)
@@ -1464,7 +1464,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may be left next to a non-empty line.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1479,7 +1479,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:12:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:8:5)
@@ -1503,7 +1503,7 @@ Hello text ## $CR user: Comment may be left next to a non-empty line.
        (path my_file.ml)
        (content
         "CR user: Let's cover the case\n  ##\n  ## Where there are empty lines that are part of the comment.\n  ## This happens when the CR has multiple paragraphs.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1524,7 +1524,7 @@ Hello ## $CR user: By the way, multiple hash is supported. The location for the 
          (stop  my_file.ml:3:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:9)
@@ -1548,7 +1548,7 @@ Hello ## $CR user: By the way, multiple hash is supported. The location for the 
        (path my_file.ml)
        (content
         "CR user: By the way, multiple hash is supported. The location for the entire\n      ## comment starts at the left-most hash character.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1591,7 +1591,7 @@ Hello text -- $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:1:30)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -1613,7 +1613,7 @@ Hello text -- $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1626,7 +1626,7 @@ Hello text -- $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:4:27)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:3)
@@ -1648,7 +1648,7 @@ Hello text -- $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\n-- span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1660,7 +1660,7 @@ Hello text -- $CR user: Comment may be left next to a non-empty line.
          (stop  my_file.ml:7:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:6:14)
@@ -1682,7 +1682,7 @@ Hello text -- $CR user: Comment may be left next to a non-empty line.
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may be left next to a non-empty line.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1718,7 +1718,7 @@ let%expect_test "single-semi-style" =
          (stop  my_file.ml:1:29)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:2)
@@ -1740,7 +1740,7 @@ let%expect_test "single-semi-style" =
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1753,7 +1753,7 @@ let%expect_test "single-semi-style" =
          (stop  my_file.ml:4:26)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:2)
@@ -1775,7 +1775,7 @@ let%expect_test "single-semi-style" =
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\n; span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1787,7 +1787,7 @@ let%expect_test "single-semi-style" =
          (stop  my_file.ml:6:76)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:6:22)
@@ -1809,7 +1809,7 @@ let%expect_test "single-semi-style" =
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may be placed after a non-empty line.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1822,7 +1822,7 @@ let%expect_test "single-semi-style" =
          (stop  my_file.ml:9:21)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:8:11)
@@ -1844,7 +1844,7 @@ let%expect_test "single-semi-style" =
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may span multiple\n         ; lines too.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1859,7 +1859,7 @@ let%expect_test "single-semi-style" =
          (stop  my_file.ml:15:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:11:4)
@@ -1883,7 +1883,7 @@ let%expect_test "single-semi-style" =
        (path my_file.ml)
        (content
         "CR user: Let's cover the case\n  ;\n  ; Where there are empty lines that are part of the comment.\n  ; This happens when the CR has multiple paragraphs.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -1919,7 +1919,7 @@ let%expect_test "double-semi-style" =
          (stop  my_file.ml:1:30)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -1941,7 +1941,7 @@ let%expect_test "double-semi-style" =
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1954,7 +1954,7 @@ let%expect_test "double-semi-style" =
          (stop  my_file.ml:4:27)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:3)
@@ -1976,7 +1976,7 @@ let%expect_test "double-semi-style" =
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\n;; span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -1988,7 +1988,7 @@ let%expect_test "double-semi-style" =
          (stop  my_file.ml:6:77)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:6:23)
@@ -2010,7 +2010,7 @@ let%expect_test "double-semi-style" =
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may be placed after a non-empty line.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -2023,7 +2023,7 @@ let%expect_test "double-semi-style" =
          (stop  my_file.ml:9:22)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:8:12)
@@ -2045,7 +2045,7 @@ let%expect_test "double-semi-style" =
      (getters (
        (path my_file.ml)
        (content "CR user: Comment may span multiple\n         ;; lines too.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -2060,7 +2060,7 @@ let%expect_test "double-semi-style" =
          (stop  my_file.ml:15:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:11:5)
@@ -2084,7 +2084,7 @@ let%expect_test "double-semi-style" =
        (path my_file.ml)
        (content
         "CR user: Let's cover the case\n  ;;\n  ;; Where there are empty lines that are part of the comment.\n  ;; This happens when the CR has multiple paragraphs.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2114,7 +2114,7 @@ let%expect_test "xml-style" =
          (stop  my_file.ml:1:36)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:5)
@@ -2136,7 +2136,7 @@ let%expect_test "xml-style" =
      (getters (
        (path    my_file.ml)
        (content "CR user: This is a comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -2149,7 +2149,7 @@ let%expect_test "xml-style" =
          (stop  my_file.ml:4:33)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:5)
@@ -2171,7 +2171,7 @@ let%expect_test "xml-style" =
      (getters (
        (path my_file.ml)
        (content "XCR user: And it can\n     span multiple lines too.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     ========================
@@ -2183,7 +2183,7 @@ let%expect_test "xml-style" =
          (stop  my_file.ml:7:50)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:7:7)
@@ -2205,7 +2205,7 @@ let%expect_test "xml-style" =
      (getters (
        (path    my_file.ml)
        (content "XCR user: What happens if it is nested?")
-       (kind    XCR)
+       (status  XCR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2242,7 +2242,7 @@ let%expect_test "nested-ml-style" =
          (stop  my_file.ml:3:74)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt XCR)
              (loc (
                (start my_file.ml:3:6)
@@ -2266,7 +2266,7 @@ let%expect_test "nested-ml-style" =
        (path my_file.ml)
        (content
         "XCR user: CR comment may be nested inside other comment in OCaml.")
-       (kind XCR)
+       (status XCR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2293,7 +2293,7 @@ let%expect_test "nested-ml-style" =
          (stop  my_file.ml:6:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -2317,7 +2317,7 @@ let%expect_test "nested-ml-style" =
        (path my_file.ml)
        (content
         "CR user: Maybe the original use case is the opposite though.\n\n    That is, a (* comment inside a CR comment. *)\n\n    We'd like to cover this case too.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2402,7 +2402,7 @@ let%expect_test "not standard" =
          (stop  my_file.ml:1:76)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:5)
@@ -2426,7 +2426,7 @@ let%expect_test "not standard" =
        (path my_file.ml)
        (content
         "CR user: This is recognized as a bash comment even though it is in what")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2445,7 +2445,7 @@ let%expect_test "not standard" =
          (stop  my_file.ml:2:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:4)
@@ -2467,7 +2467,7 @@ let%expect_test "not standard" =
      (getters (
        (path    my_file.ml)
        (content "CR user1: A CR in a odoc comment.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2483,7 +2483,7 @@ let%expect_test "not standard" =
          (stop  my_file.ml:2:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:1:3)
@@ -2505,7 +2505,7 @@ let%expect_test "not standard" =
      (getters (
        (path    my_file.ml)
        (content "CR user1: A CR with a tab separator.")
-       (kind    CR)
+       (status  CR)
        (qualifier (None))
        (priority Now))))
     |}];
@@ -2525,7 +2525,7 @@ let%expect_test "not standard" =
          (stop  my_file.ml:3:0)))
        (header (
          Ok (
-           (kind (
+           (status (
              (txt CR)
              (loc (
                (start my_file.ml:2:5)
@@ -2547,7 +2547,7 @@ let%expect_test "not standard" =
      (getters (
        (path my_file.ml)
        (content "CR user1: A CR with spaces and newline separators.")
-       (kind CR)
+       (status CR)
        (qualifier (None))
        (priority Now))))
     |}];